refactor: rename common file to add .js
