+++ packages/client/src/pages/Algorithm/Common/Canimate.js
